CSQ_project_v5/requirements.txt
streamlit
numpy
PyPDF2
python-docx
chromadb
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain-google-genai
pydantic
chunking_evaluation @ git+https://github.com/brandonstarxel/chunking_evaluation.git@d451fc4cf56e417b755994b4ca5212fd5057c0d2
